refactor(frontend): クイズ大会モードの管理者側ロジックをApp.jsxからhookへ切り出す
設計:
- 選定内容: App.jsxに直接展開されていたクイズ大会モードの管理者側ロジック（ルーム作成・開設中ルーム一覧の取得・WebSocket接続・早押し判定・状態ブロードキャスト）をhooks/useQuizRoomAdmin.jsへ抽出した。対応するテスト（App.test.jsxの末尾にあった16件）もhooks/useQuizRoomAdmin.test.jsxへ移動した。
- 却下内容: 読み上げ・札めくり進行（playKarutaInternal等）や結果画面演出の抽出は本PRには含めない。
- 理由: issue #607（App.jsx 1,820行・App.test.jsx 3,245行の再肥大化）自身が「一括リファクタリングではなく抽出単位ごとに小さなPRへ分ける」段階的な進め方を推奨しているため、まず最も自己完結的なクイズ大会管理者ロジックのみを切り出した。残る2つの抽出単位（読み上げ進行のhook化、結果画面・演出のコンポーネント化）はリスクが高い・変更範囲が広いため別PRとし、issue #607に進捗をコメントして開いたままにする。

影響:
- 影響モジュール: frontend/src/App.jsx, frontend/src/hooks/useQuizRoomAdmin.js（新規）, frontend/src/App.test.jsx, frontend/src/hooks/useQuizRoomAdmin.test.jsx（新規）
- 振る舞いの変更: なし。ロジック自体はコメントを含めて一切変更せず、既存の関数・stateをhookへ移動しただけ（App.jsx側からの見え方・戻り値の名前も変更していない）。

テスト:
- 追加済み: 既存の16件のテストをhooks/useQuizRoomAdmin.test.jsxへ移動し、そのまま（<App />を実際にレンダーする既存の統合テスト形式のまま）残した。
- 未追加: 新規のロジック追加ではなく既存コードの移動のため、新規テストケースの追加は行っていない。npm run lint、npx vitest run（全172件）がともに成功することを確認済み。
